Refactor : jwt, stomp 폴더 구조 변경
