+++ src/main/java/leaguehub/leaguehubbackend/config/SecurityConfiguration.java
@@ -1,5 +1,7 @@
 package leaguehub.leaguehubbackend.config;
 
+import leaguehub.leaguehubbackend.config.jwt.JwtAuthenticationEntryPoint;
+import leaguehub.leaguehubbackend.config.jwt.JwtAuthenticationProcessingFilter;
 import leaguehub.leaguehubbackend.repository.member.MemberRepository;
 import leaguehub.leaguehubbackend.service.jwt.JwtService;
 import lombok.RequiredArgsConstructor;
@@ -13,7 +15,6 @@ import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-import org.springframework.web.servlet.handler.HandlerMappingIntrospector;
 
 @Configuration
 @EnableWebSecurity
