+++ src/test/java/leaguehub/leaguehubbackend/config/JwtAuthenticationProcessingFilterTest.java
@@ -4,6 +4,7 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import leaguehub.leaguehubbackend.config.jwt.JwtAuthenticationProcessingFilter;
 import leaguehub.leaguehubbackend.exception.auth.exception.AuthExpiredTokenException;
 import leaguehub.leaguehubbackend.exception.auth.exception.AuthInvalidTokenException;
 import leaguehub.leaguehubbackend.exception.auth.exception.AuthMemberNotFoundException;
